refactor(core): replace inline token estimation with estimate_tokens(), extract remove_system_messages helpers (#2778, #2779)
Replace ~15 inline `len() / 4` byte-based token estimations with
`zeph_common::text::estimate_tokens()` (chars-based, more accurate for
non-ASCII). Extract two private helpers in ContextAssembly —
`remove_by_prefix` and `remove_by_part_or_prefix` — reducing 11
duplicated retain blocks to one-liner delegates.

Part of #2776.

+++ crates/zeph-core/src/agent/persistence.rs
@@ -3627,10 +3627,10 @@ mod tests {
 
     #[test]
     fn trajectory_extraction_slice_handles_few_messages() {
-        let max_messages: u64 = 20;
+        let max_messages: usize = 20;
         let total_messages = 5usize;
 
-        let tail_start = total_messages.saturating_sub(max_messages as usize);
+        let tail_start = total_messages.saturating_sub(max_messages);
         let window = total_messages - tail_start;
 
         assert_eq!(window, 5, "should return all messages when fewer than max");
